requirements.txt
Django==5.2.9
djangorestframework==3.16.1
psycopg[binary]==3.3.2
